requirements.txt
# =============================================================================
# LABBAIK AI v6.0 - Dependencies
# =============================================================================

# Core framework
streamlit>=1.28.0

# AI Services
groq>=0.4.0
openai>=1.3.0

# Database
psycopg2-binary>=2.9.9

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2

# Optional: RAG capabilities
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.0

# Email validation
email-validator>=2.0.0

httpx
